[pulz-ecs] update len() to usize, include macros in workspace

+++ crates/ecs/src/entity.rs
@@ -51,7 +51,7 @@ impl Entities {
     }
 
     #[inline]
-    pub fn len(&self) -> u32 {
+    pub fn len(&self) -> usize {
         self.arena.len()
     }
